module_api/requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
coverage==4.4.1
feedparser==5.2.1
Flask==0.12.2
GoogleAppEngineCloudStorageClient==1.9.22.1
google-api-python-client==1.6.2
mock==1.3.0
protobuf==2.6.1
python-twitter==3.3
pytz==2017.2
requests-toolbelt==0.8.0
sendgrid==4.2.0
urllib3==1.21.1
user_agents==1.1.0
